pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gauntlet"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["requests", "python-dotenv"]

[tool.setuptools]
packages = ["gauntlet"]
